Switch to using better files and centroids
