+++ switch_model/tools/graph/main.py
@@ -6,7 +6,6 @@ See docs/Graphs.md to learn how to add graphs.
 # Standard packages
 import functools
 import importlib
-import traceback
 import os
 import warnings
 from typing import List, Dict, Optional
@@ -15,12 +14,13 @@ from typing import List, Dict, Optional
 import numpy as np
 import pandas as pd
 from PIL import Image
-from matplotlib import pyplot as plt, colors
+from matplotlib import pyplot as plt
 import seaborn as sns
 import matplotlib
 import plotnine
 
 # Local imports
+from switch_model.tools.graph.maps import GraphMapTools
 from switch_model.utilities import StepTimer, get_module_list, query_yes_no, catch_exceptions
 
 # When True exceptions that are thrown while graphing will be caught
@@ -107,232 +107,6 @@ class Scenario:
 
     def __exit__(self, exc_type, exc_val, exc_tb):
         os.chdir(Scenario.root_path)
-
-
-class GraphMapTools:
-    def __init__(self, graph_tools):
-        self._tools = graph_tools
-        self._wecc_states = None
-        self._center_points = None
-        self._shapely = None
-        self._geopandas = None
-
-    def _load_maps(self):
-        """
-        Plots the WECC states as a background on the provided axes
-        and returns a dataframe with a mapping of load zones to center points.
-        """
-        if self._wecc_states is None or self._center_points is None:
-            try:
-                import geopandas
-            except ModuleNotFoundError:
-                raise Exception(
-                    "Could not find package 'geopandas'. If on Windows make sure you install it through conda.")
-
-            import shapely
-            self._shapely = shapely
-            self._geopandas = geopandas
-
-            # Read shape files
-            try:
-                wecc_lz = geopandas.read_file(self._tools.get_file_path("maps/wecc_102009.shp", from_inputs=True),
-                                              crs="ESRI:102009")
-                self._wecc_states = geopandas.read_file(
-                    self._tools.get_file_path("maps/wecc_states_4326.shp", from_inputs=True))
-            except FileNotFoundError:
-                raise Exception("Can't create maps, shape files are missing. Try running switch get_inputs.")
-
-            wecc_lz = wecc_lz.rename({"LOAD_AREA": "gen_load_zone"}, axis=1)
-
-            # New dataframe with the centerpoint geometry
-            self._center_points = geopandas.GeoDataFrame(
-                wecc_lz[["gen_load_zone"]],
-                geometry=wecc_lz["geometry"].centroid,
-                crs="ESRI:102009",
-            ).to_crs("EPSG:4326")
-
-        return self._wecc_states, self._center_points
-
-    def _plot_states(self, ax):
-        states, _ = self._load_maps()
-        states.plot(ax=ax, lw=0.5, edgecolor="white", color="#E5E5E5", zorder=-999)
-        ax.axis("off")
-
-    @staticmethod
-    def _pie_plot(x, y, ratios, colors, size, ax=None):
-        # REFERENC: https://tinyurl.com/app/myurls
-        # determine arches
-        start = 0.0
-        xy = []
-        s = []
-        for ratio in ratios:
-            x0 = [0] + np.cos(
-                np.linspace(2 * np.pi * start, 2 * np.pi * (start + ratio), 30)
-            ).tolist()  # 30
-            y0 = [0] + np.sin(
-                np.linspace(2 * np.pi * start, 2 * np.pi * (start + ratio), 30)
-            ).tolist()  # 30
-
-            xy1 = np.column_stack([x0, y0])
-            s1 = np.abs(xy1).max()
-
-            xy.append(xy1)
-            s.append(s1)
-            start += ratio
-
-        for xyi, si, c in zip(xy, s, colors):
-            ax.scatter(
-                [x], [y], marker=xyi, s=size * si ** 2, c=c, edgecolor="k"
-            )
-
-    def graph_pie_chart(self, df, max_size=2500, fixed_size=None, ax=None):
-        """
-        Graphs the data from the dataframe to a map pie chart.
-        The dataframe should have 3 columns, gen_load_zone, gen_type and value.
-        """
-        _, center_points = self._load_maps()
-
-        # Scale the dataframe so the pie charts have the right size
-        current_max_size = df.groupby("gen_load_zone")["value"].sum().max()
-        df["value"] *= max_size / current_max_size
-
-        if ax is None:
-            ax = self._tools.get_axes()
-        self._plot_states(ax)
-        df = df.merge(center_points, on="gen_load_zone")
-
-        assert not df["gen_type"].isnull().values.any()
-        colors = self._tools.get_colors()
-        for index, group in df.groupby(["gen_load_zone"]):
-            x, y = group["geometry"].iloc[0].x, group["geometry"].iloc[0].y
-            group_sum = group.groupby("gen_type")["value"].sum().sort_values()
-            group_sum = group_sum[group_sum != 0].copy()
-
-            tech_color = [colors[tech] for tech in group_sum.index.values]
-            total_size = group_sum.sum()
-            ratios = (group_sum / total_size).values
-            if fixed_size is not None:
-                total_size = fixed_size
-            GraphMapTools._pie_plot(x, y, ratios, tech_color, ax=ax, size=total_size)
-        return ax
-
-    def _bin_data(self, data, bins, mapping):
-        """
-        Puts the data into bins and returns the binned data as well as labels and data needed
-        to make a legend.
-        data : Series containing the data to transform
-        bins : list containing the cuts between the bins (not including the ends)
-        mapping : function that given the bin index returns the desired value
-        Returns a tuple with 3 elements. The first is an array with the binned data.
-        The second is a list of labels for the bins
-        The third is the value associated with each bin matching the label ordering.
-        """
-        binned_data = []
-        num_bins = len(bins) + 1
-        for point in data:
-            if point > bins[-1]:
-                binned_data.append(num_bins - 1)
-                continue
-            for i, bin in enumerate(bins):
-                if point <= bin:
-                    binned_data.append(i)
-                    break
-        binned_and_mapped = []
-        for point in binned_data:
-            binned_and_mapped.append(mapping(point))
-
-        labels = [f"<{bins[0]}"]
-        legend_vals = [mapping(0.0)]
-        for i in range(1, len(bins)):
-            legend_vals.append(mapping(i))
-            labels.append(f"{bins[i - 1]}-{bins[i]}")
-        labels.append(f">{bins[-1]}")
-        legend_vals.append(mapping(num_bins - 1))
-
-        return binned_and_mapped, labels, legend_vals
-
-    def graph_squares(self, df, bins=(5, 8, 10, 15), cmap="viridis", ax=None, size=20, unit=" h"):
-        """
-        Graphs the data from the dataframe to a squares on each cell.
-        The dataframe should have 2 columns, gen_load_zone and value.
-        """
-        _, center_points = self._load_maps()
-
-        cmap_func = cmap
-        num_bins = len(bins) + 1
-        if type(cmap_func) == str:
-            cmap_func = plt.get_cmap(cmap_func)
-
-        if ax is None:
-            ax = self._tools.get_axes()
-        self._plot_states(ax)
-        df = df.merge(center_points, on="gen_load_zone", validate="one_to_one")
-        colors, legend_labels, legend_colors = self._bin_data(df["value"], bins,
-                                                              lambda x: cmap_func(float(x / (num_bins - 1))))
-        for i, row in df.iterrows():
-            x, y = row["geometry"].x, row["geometry"].y
-            ax.scatter(x, y, marker="s", s=size * 2, color="dimgray")  # Add a black border
-            ax.scatter(x, y, marker="s", s=size, color=colors[i])
-        ax.legend(
-            title="Storage duration (h)",
-            handles=[self._tools.plt.lines.Line2D([], [], color=c, marker="s", markersize=5, label=l, linestyle="None")
-                     for c, l in zip(legend_colors, legend_labels)])
-
-    def graph_points(self, df, ax=None):
-        """
-        Graphs a point in each load zone based on a dataframe with two columns
-        - gen_load_zone
-        - value
-        """
-        _, center_points = self._load_maps()
-
-        df = df.merge(center_points, on="gen_load_zone")
-        # Cast to GeoDataFrame
-        df = self._geopandas.GeoDataFrame(df[["geometry", "value"]], geometry="geometry")
-
-        if ax is None:
-            ax = self._tools.get_axes()
-        self._plot_states(ax)
-        df.plot(ax=ax, column="value", legend=True, cmap="coolwarm", markersize=30, norm=colors.CenteredNorm())
-
-    def graph_transmission(self, df, cutoff, ax=None, legend=True, zorder=-50):
-        """
-        Graphs the data frame a dataframe onto a map.
-        The dataframe should have 4 columns:
-        - from: the load zone where we're starting from
-        - to: the load zone where we're going to
-        - value: the value to plot
-        """
-        if ax is None:
-            ax = self._tools.get_axes()
-        _, center_points = self._load_maps()
-
-        # Merge duplicate rows if table was unidirectional
-        df[["from", "to"]] = df[["from", "to"]].apply(sorted, axis=1, result_type="expand")
-        df = df.groupby(["from", "to"], as_index=False)["value"].sum()
-
-        df = df.merge(
-            center_points.add_prefix("from_"),
-            left_on="from",
-            right_on="from_gen_load_zone",
-        ).merge(
-            center_points.add_prefix("to_"),
-            left_on="to",
-            right_on="to_gen_load_zone"
-        )[["from_geometry", "to_geometry", "value"]]
-
-        from shapely.geometry import LineString
-
-        def make_line(r):
-            return LineString([r["from_geometry"], r["to_geometry"]])
-
-        df["geometry"] = df.apply(make_line, axis=1)
-        # Cast to GeoDataFrame
-        df = self._geopandas.GeoDataFrame(df[["geometry", "value"]], geometry="geometry")
-
-        self._plot_states(ax)
-        df.plot(ax=ax, column="value", legend=legend, cmap="Reds", zorder=zorder, norm=colors.LogNorm(vmin=cutoff, vmax=df.value.max()))
-        return ax
 
 
 class TransformTools:
