--- switch_model/wecc/get_inputs/graph_config/maps/wecc_102009.qmd
@@ -1,43 +1,0 @@
-<!DOCTYPE qgis PUBLIC 'http://mrcc.com/qgis.dtd' 'SYSTEM'>
-<qgis version="3.20.0-Odense">
-  <identifier></identifier>
-  <parentidentifier></parentidentifier>
-  <language></language>
-  <type>dataset</type>
-  <title></title>
-  <abstract></abstract>
-  <contact>
-    <name></name>
-    <organization></organization>
-    <position></position>
-    <voice></voice>
-    <fax></fax>
-    <email></email>
-    <role></role>
-  </contact>
-  <links/>
-  <fees></fees>
-  <encoding></encoding>
-  <crs>
-    <spatialrefsys>
-      <wkt></wkt>
-      <proj4></proj4>
-      <srsid>0</srsid>
-      <srid>0</srid>
-      <authid></authid>
-      <description></description>
-      <projectionacronym></projectionacronym>
-      <ellipsoidacronym></ellipsoidacronym>
-      <geographicflag>false</geographicflag>
-    </spatialrefsys>
-  </crs>
-  <extent>
-    <spatial crs="" dimensions="2" minz="0" miny="0" maxy="0" minx="0" maxx="0" maxz="0"/>
-    <temporal>
-      <period>
-        <start></start>
-        <end></end>
-      </period>
-    </temporal>
-  </extent>
-</qgis>
