--- switch_model/wecc/get_inputs/graph_config/maps/wecc_states_4326.cpg
@@ -1,1 +1,0 @@
-UTF-8
